+++ packages/tpstech-catppuccin-vscode/src/compile.ts
@@ -1,6 +1,6 @@
 import { FlavorName } from "@catppuccin/palette";
-import { compileTheme, defaultOptions } from "tsdevau-catppuccin-vsc/src/theme";
-import { ThemeOptions } from "tsdevau-catppuccin-vsc/src/types";
+import { compileTheme, defaultOptions } from "tpstech-catppuccin-vsc/src/theme";
+import { ThemeOptions } from "tpstech-catppuccin-vsc/src/types";
 
 /**
  * Compiles the Catppuccin for VSCode theme with optional custom overrides.
@@ -38,6 +38,6 @@ export const compile = (
   // like in the original file, we need to shim the name for the Shiki theme
   return {
     ...compiled,
-    name: `tsdevau-catppuccin-${flavor}`,
+    name: `tpstech-catppuccin-${flavor}`,
   };
 };