+++ packages/tpstech-catppuccin-vsc/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [3.17.1](https://github.com/tsdevau/catppuccin-vscode/compare/tsdevau-catppuccin-vsc-v3.17.0...tsdevau-catppuccin-vsc-v3.17.1) (2025-03-30)
+## [3.17.1](https://github.com/tsdevau/catppuccin-vscode/compare/tpstech-catppuccin-vsc-v3.17.0...tpstech-catppuccin-vsc-v3.17.1) (2025-03-30)
 
 
 ### Bug Fixes
@@ -370,7 +370,7 @@ redesign has grown on you, and that the theme has improved since September.
 
 ### Features
 
-* **options:** support hex codes in `tsdevau-catppuccin.customUIColors` ([#126](https://github.com/catppuccin/vscode/issues/126)) ([6df739f](https://github.com/catppuccin/vscode/commit/6df739f9c7dce789216f4afd76df21e3fb3abb14)), closes [#109](https://github.com/catppuccin/vscode/issues/109)
+* **options:** support hex codes in `tpstech-catppuccin.customUIColors` ([#126](https://github.com/catppuccin/vscode/issues/126)) ([6df739f](https://github.com/catppuccin/vscode/commit/6df739f9c7dce789216f4afd76df21e3fb3abb14)), closes [#109](https://github.com/catppuccin/vscode/issues/109)
 * **ui:** consistently use peach for warnings & yellow for modified ([22ee5cf](https://github.com/catppuccin/vscode/commit/22ee5cfb422124aa5a9368d26dea58145fbe0cd3))
 * **ui:** rework minimap colors ([#129](https://github.com/catppuccin/vscode/issues/129)) ([a9c0e0d](https://github.com/catppuccin/vscode/commit/a9c0e0d1fee9925ee904da033a2e228f80e583f1))
 * **ui:** use accent color for more focused elements ([#128](https://github.com/catppuccin/vscode/issues/128)) ([2b1a8a7](https://github.com/catppuccin/vscode/commit/2b1a8a7d9a0c687519ec72e4b333ecbe71c43f09)), closes [#91](https://github.com/catppuccin/vscode/issues/91)